% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_amazon_cn.R
\name{add_amazon_cn}
\alias{add_amazon_cn}
\title{Create BibTex from an Amazon.cn book}
\usage{
add_amazon_cn(url, is_paste = TRUE)
}
\arguments{
\item{url}{URL.}

\item{is_paste}{Logical, by default \code{TRUE}}
}
\description{
Create BibTex from an Amazon.cn book
}
