% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibtex.R
\name{add_bibtex}
\alias{add_bibtex}
\title{Bibtex and Biblatex Template}
\usage{
add_bibtex(type = "more")
}
\arguments{
\item{type}{Character, by default \code{"more"}}
}
\value{
Character.
}
\description{
This function helps users to add Add a bib(La)tex template
}
\examples{
add_bibtex("online")
add_bibtex("more")
add_bibtex("book")
add_bibtex("manual")
}
\author{
Jiaxiang Li
}
