% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tail_refs.R
\name{add_tail_refs}
\alias{add_tail_refs}
\title{Add reference tag into quote paragraphs.}
\usage{

  add_tail_refs(path = "jiangning_weaving/analysis/para_length_5_notes.Rmd",
  append_text = "[@lastname2020")
}
\arguments{
\item{path}{The RMarkdown document path.}

\item{append_text}{The BibTex alias.}
}
\description{
Add reference tag into quote paragraphs.
}
