% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibtex.R
\name{add_zhihu}
\alias{add_zhihu}
\title{Add BibTex for a Zhihu URL.}
\usage{
add_zhihu(input = "", is_paste = TRUE)
}
\arguments{
\item{input}{The Zhihu shared text.}

\item{is_paste}{Logical, by default \code{TRUE}}
}
\value{
Character.
}
\description{
Add BibTex for a Zhihu URL.
}
\examples{
\donttest{add_zhihu("url")}
}
