% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibtex.R
\name{add_bibtex}
\alias{add_bibtex}
\title{BibTex and BibLaTex template}
\usage{
add_bibtex(type = "more")
}
\value{
Character.
}
\description{
This function helps users to add Add a bib(La)tex template
}
\author{
Jiaxiang Li
}
