% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibtex.R
\name{add_bibtex}
\alias{add_bibtex}
\title{Bibtex and Biblatex Template}
\usage{
add_bibtex(type = "more", is_paste = TRUE)
}
\arguments{
\item{type}{Character, by default \code{"more"}}

\item{is_paste}{Logical, by default \code{TRUE}}
}
\value{
Character.
}
\description{
This function helps users to add Add a bib(La)tex template. It is pasted on the clipboard.
}
\examples{
add_bibtex("online", is_paste = FALSE)
add_bibtex("book", is_paste = FALSE)
add_bibtex("manual", is_paste = FALSE)
add_bibtex("more")
}
\author{
Jiaxiang Li
}
