% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibtex.R
\name{add_datacamp}
\alias{add_datacamp}
\title{Add BibTex for a DataCamp URL.}
\usage{
add_datacamp(url = "")
}
\arguments{
\item{url}{URL.}
}
\description{
Add BibTex for a DataCamp URL.
}
\examples{
add_datacamp("https://www.datacamp.com/courses/extreme-gradient-boosting-with-xgboost")
}
