% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_bibtex.R
\name{add_kaggle}
\alias{add_kaggle}
\title{Add BibTex for a Kaggle URL.}
\usage{
add_kaggle(url = "")
}
\arguments{
\item{url}{URL.}
}
\description{
Add BibTex for a Kaggle URL.
}
\examples{
add_kaggle("https://www.kaggle.com/lijiaxiang/stacking")
}
