% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_wechat.R
\name{add_wechat}
\alias{add_wechat}
\title{Add BibTex for a WeChat article URL.}
\usage{
add_wechat(url = "", is_paste = TRUE)
}
\arguments{
\item{url}{URL.}

\item{is_paste}{Logical, by default \code{TRUE}}
}
\description{
Add BibTex for a WeChat article URL.
}
\examples{
add_wechat("https://mp.weixin.qq.com/s/2-1taZ5o4uzVcGMe5u4P-A")
}
